// Sipro/Sipro/ClientApp/src/app/components/proyecto/model/model.proyecto.ts
export class Proyecto{
    id: number;
    nombre: string;
    objetivo: string;
    descripcion: string;
    snip: number;
    proyectotipoid: number;
    proyectotipo: string;
    unidadejecutora: string;
    unidadejecutoraid: number;
    entidadentidad: number;
    entidadnombre: string;
    ejercicio: number;
    fechaCreacion: string;
    usuarioCreo: string;
    fechaactualizacion: string;
    usuarioactualizo: string;
    programa?: number;
    subprograma?: number;
    proyecto?: number;
    actividad?: number;
    obra?: number;
    renglon?: number;
    ubicacionGeografica?: number;
    longitud: string;
    latitud: string;
    directorProyectoId: number;
    directorProyectoNmbre: string;
    costo: number;
    acumulacionCosto: number;
    acumulacionCostoNombre: string;
    objetivoEspecifico: string;
    visionGeneral: string;
    ejecucionFisicaReal: number;
    proyectoClase: number;
    projectCargado: number;
    prestamoId: number;
    fechaInicio: Date;
    fechaFin: Date;
    observaciones: string;
    fechaInicioReal: Date;
    fechaFinReal: Date;
    congelado: number;
    coordinador: number;
    porcentajeAvance: number;
    permisoEditarCongelar: boolean;
    lineaBaseId: number;
}